[FSDP][3/N] Split `fully_shard` unit tests
[ghstack-poisoned]
